third_party/WebKit/LayoutTests/fast/dom/shadow/apply-deep-in-document-scope.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<style>div /deep/ #inner { color: green }</style>
<div>
    <div id="inner">This text should be green</div>
</div>
<script>
    description("/deep/ as a descendant selector in document without shadow trees.");

    shouldBeEqualToString("getComputedStyle(inner).color", "rgb(0, 128, 0)");
</script>
