third_party/WebKit/LayoutTests/fast/dom/shadow/apply-deep-in-document-scope-expected.txt

CONSOLE WARNING: line 1: /deep/ combinator in CSS is deprecated and will be removed in M63, around December 2017. See https://www.chromestatus.com/features/4964279606312960 for more details.
/deep/ as a descendant selector in document without shadow trees.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(inner).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
This text should be green
